+++ demo/utils/constants.js
@@ -157,6 +157,11 @@ export const componentList = [
         path: '/pages/Selector/index',
       },
       {
+        name: 'Slider',
+        nameZN: '滑动输入条',
+        path: '/pages/Slider/index',
+      },
+      {
         name: 'Stepper',
         nameZN: '步进器',
         path: '/pages/Stepper/index',
@@ -195,6 +200,11 @@ export const componentList = [
         name: 'Dialog',
         nameZN: '对话框',
         path: '/pages/Dialog/index',
+      },
+      {
+        name: 'Empty',
+        nameZN: '空状态',
+        path: '/pages/Empty/index',
       },
       {
         name: 'Loading',