+++ src/VideoUpload/index.ts
@@ -5,27 +5,51 @@ import { chooseVideo, uploadFile } from '../_util/promisify';
 Component({
   props: VideoUploadDefaultProps,
   data: {
+    id: '',
     fileList: [],
     playVideoUrl: '',
   } as IVideoUploadData,
-  didMount() {
-    const { value, id } = this.props;
 
+  didMount() {
     // eslint-disable-next-line @typescript-eslint/ban-ts-comment
     // @ts-ignore
-    this.videoContext = my.createVideoContext(id || 'myVideo');
+    this.videoContext = my.createVideoContext(`video_${this.$page.$id}_${this.$id}`);
     this.setData({
-      fileList: value
-    });
+      id: `video_${this.$page.$id}_${this.$id}`
+    })
+    this.handleValue();
   },
+
   didUpdate(prevProps) {
     if (!equal(prevProps.value, this.props.value)) {
-      this.setData({
-        fileList: this.props.value
-      })
+      this.handleValue();
     }
   },
+
   methods: {
+    handleValue() {
+      let curValue;
+      const { value } = this.props;
+
+      if (typeof value === 'string') {
+        curValue = [].concat(value);
+      } else if (
+        Array.isArray(value) &&
+        value.length &&
+        value.some(v => typeof v === 'string')
+      ) {
+        curValue = value.map(v => (typeof v === 'string' ? {
+          url: v,
+          status: 'done'
+        } : v));
+      } else {
+        curValue = value;
+      }
+      this.setData({
+        fileList: curValue
+      });
+    },
+
     async onChooseVideo() {
       const { camera, maxDuration, sourceType } = this.props;
 
@@ -44,6 +68,7 @@ Component({
         })
       }
     },
+
     async uploadFile(file) {
       const { action, filename, formData, onBeforeUpload, onUpload, onAfterUpload } = this.props;
       const { fileList } = this.data;
@@ -121,6 +146,7 @@ Component({
         });
       }
     },
+
     updateFileList(path, status, url?) {
       const { fileList } = this.data;
       const { onChange } = this.props;
@@ -142,6 +168,7 @@ Component({
 
       onChange && onChange.call(this.props, tempFileList);
     },
+
     async onDeleteVideo(e) {
       const { fileList } = this.data;
       const { onDelete, onChange } = this.props;
@@ -158,6 +185,7 @@ Component({
       });
       onChange && onChange.call(this.props, tempFileList);
     },
+
     onPreviewVideo(e) {
       const { previewVideoUrl } = e.target.dataset;
 
@@ -168,11 +196,13 @@ Component({
       });
 
     },
+
     onPlay() {
       this.videoContext.requestFullScreen({
         direction: 0
       });
     },
+
     onFullScreenChange(e) {
       if (!e.detail.fullScreen) {
         this.videoContext.pause();